Give access to underlying editText view
Hi this is more of a suggestion than a full blown PR but I started to use NumberSlidingPicker and I found myself wanting to use input mask and or other extension for EditText (eg https://github.com/RedMadRobot/input-mask-android). Most of them require a either reference to the EditText or extend it through inheritance and changing `editText` visibility help for the first case.
I don't have a lot of experience in Java or Kotlin so there might be issues or side effects with my solution.